% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_log}
\alias{filter_log}
\title{A logging function that captures the shiny namespace}
\usage{
filter_log(..., ns, verbose = TRUE)
}
\arguments{
\item{...}{fields to be logged}

\item{ns}{the name of the current active namespace. if missing, will be
derived from parent environment.}

\item{verbose}{whether log should be written}
}
\value{
\code{NULL}
}
\description{
A logging function that captures the shiny namespace
}
