% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_log_ns_style}
\alias{filter_log_ns_style}
\title{A crayon style derived from a rough hash of the namespace name}
\usage{
filter_log_ns_style(txt)
}
\arguments{
\item{txt}{ns name to format}
}
\description{
A crayon style derived from a rough hash of the namespace name
}
