% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_log_style}
\alias{filter_log_style}
\title{A crayon style for the filter log}
\usage{
filter_log_style(...)
}
\arguments{
\item{...}{passed to a \link{crayon} crayon object}
}
\description{
A crayon style for the filter log
}
