% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter.R
\name{get_dataFilter_class}
\alias{get_dataFilter_class}
\title{Return the class used for dispatch to shiny_vector_filter formatted using
pillar}
\usage{
get_dataFilter_class(obj)
}
\arguments{
\item{obj}{the object whose class for dispatch should be derived}
}
\value{
a pillar formatted class name
}
\description{
Return the class used for dispatch to shiny_vector_filter formatted using
pillar
}
