% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullor}
\alias{nullor}
\alias{\%||\%}
\title{Shorthand operator for NULL fallback}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{lhs for null-or operation}

\item{rhs}{rhs for null-or operation}
}
\description{
Shorthand operator for NULL fallback
}
\examples{
`\%||\%` <- shinyDataFilter:::`\%||\%`

NULL \%||\% 1
# [1] 1

1 \%||\% 2
# [1] 1

}
