% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_data_filter_item.R
\name{shiny_data_filter_item_ui}
\alias{shiny_data_filter_item_ui}
\title{A single filter item as part of a data filter module panel}
\usage{
shiny_data_filter_item_ui(inputId, verbose = FALSE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{verbose}{a \code{logical} value indicating whether or not to print log
statements out to the console}
}
\value{
a shiny \code{\link[shiny]{wellPanel}} to house the filter
}
\description{
This sub-module contains the logic for the filter item iteself. This contains
general ui elements such as the button to remove the filter and the grip for
drag-and-drop. It also handles the conditional logic of showing a column
selector or the column name. When a column is selected it also embeds another
module showing module for filtering the specified vector of data.
}
