% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter_factor_many.R
\name{shiny_vector_filter_factor_many}
\alias{shiny_vector_filter_factor_many}
\title{A vector filter for factors with only a few choices}
\usage{
shiny_vector_filter_factor_many(input, output, session,
  x = shiny::reactive(factor()), filter_na = shiny::reactive(FALSE),
  verbose = FALSE)
}
\arguments{
\item{input}{requisite shiny module field specifying incoming ui input
reactiveValues}

\item{output}{requisite shiny module field capturing output for the shiny
data filter ui}

\item{session}{requisite shiny module field containing the active shiny
session}

\item{x}{a reactive expression resolving to the vector to filter}

\item{filter_na}{a logical value indicating whether to filter \code{NA}
values from the \code{x} vector}

\item{verbose}{a \code{logical} value indicating whether or not to print log
statements out to the console}
}
\value{
a \code{\link[shiny]{reactiveValues}} list containing a logical
  vector called "mask" which can be used to filter the provided vector and an
  element "code" which is the expression used to generate the mask.
}
\description{
A vector filter for factors with only a few choices
}
