% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter.R
\name{shiny_vector_filter_ui}
\alias{shiny_vector_filter_ui}
\title{A stub UI for a vector filter module}
\usage{
shiny_vector_filter_ui(data, inputId)
}
\arguments{
\item{data}{the data object to be filtered}

\item{inputId}{The \code{input} slot that will be used to access the value.}
}
\value{
an empty \code{link[shiny]{div}} or a \code{link[shiny]{uiOutput}}
  placeholder expecting output "ui"
}
\description{
This is a stub of a UI which assumes any dispatched module will create the
needed UI in the "ui" uiOutput target
}
