% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinytest_path}
\alias{shinytest_path}
\title{Helper to debug shinytests so they work interactively as well as during test}
\usage{
shinytest_path(path)
}
\arguments{
\item{path}{a path within the tests/shinytest/ directory}
}
\value{
a path that works irrespective of how the code is executed
}
\description{
Helper to debug shinytests so they work interactively as well as during test
}
