% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_leading_ws}
\alias{strip_leading_ws}
\title{Strip leading white space from a block of text}
\usage{
strip_leading_ws(txt, simplify = TRUE)
}
\arguments{
\item{txt}{text to strip leading whitespace from}

\item{simplify}{whether to simplify down to a character vector}
}
\value{
the block of text with entire columns of leading whitespace removed
}
\description{
Strip leading white space from a block of text
}
